Added: Extension for topk patches
Extension of the project added that tries to improve the AUROC test results by selecting only a few patches of the image to compare distance instead of all of them. The intuition is that images could have parts where generation is more obvious or concentrated. This includes changes to the main and a few functions such as score loaders for both Noise/Blur.